backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Scraping
playwright==1.41.1
beautifulsoup4==4.12.3
trafilatura==1.6.4
requests==2.31.0
httpx==0.26.0

# PDF Processing
PyMuPDF==1.23.22
Pillow==10.2.0

# OCR
google-cloud-documentai==2.26.0
pytesseract==0.3.10

# Arabic NLP
camel-tools==1.5.7
pyarabic==0.6.15

# AI & Embeddings
openai==1.12.0
tiktoken==0.6.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1
validators==0.22.0

# Date conversion (Hijri to Gregorian)
hijri-converter==2.3.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# CORS
fastapi-cors==0.0.6
